requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
firebase-admin
google-cloud-firestore
openai
pinecone
python-dotenv